backend/requirements.txt
# Krishi-Net Backend Dependencies
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-multipart==0.0.9
aiofiles==23.2.1
httpx==0.26.0
Pillow==10.2.0
numpy==1.26.4
tensorflow==2.15.0
twilio==8.13.0
alembic==1.13.1
